Implemented the game

+++ src/common.rs
@@ -7,11 +7,11 @@
  *
  */
 
-#[derive(Serialize, Deserialize, Debug)]
+#[derive(Serialize, Deserialize, Debug, PartialEq, Eq, Hash, Clone)]
 pub struct Point
 {
-    pub x: u32,
-    pub y: u32,
+    pub x: usize,
+    pub y: usize,
 }
 
 #[derive(Serialize, Deserialize, Debug)]
